Push props from EnvironmentList to EnvironmentItem
Because they were not needed in the list, we can create them later,
sparing us from passing them on.
